# When building a tree for package namespaces configuration remove couple of properties to reduce memory allocations.