--- a/src/NuGet.Core/NuGet.Configuration/PackageNamespaces/SearchTree.cs
+++ b/src/NuGet.Core/NuGet.Configuration/PackageNamespaces/SearchTree.cs
@@ -115,7 +115,12 @@ namespace NuGet.Configuration
                 }
             }
 
-            return currentNode.PackageSources ?? (longestMatchingPrefixNode?.PackageSources);
+            if (i == term.Length && currentNode.PackageSources != null)
+            {
+                return currentNode.PackageSources;
+            }
+
+            return longestMatchingPrefixNode?.PackageSources;
         }
     }
 }